requirements.txt
#
# This file contains all requirements and is meant for dependency analyzers etc.
# Do not use this file to install requirements, use one of the "requirements/*.txt" files instead.
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    manage/deps-upgrade.sh
#
asn1crypto==0.24.0        # via cryptography
bandit==1.6.1             # via flake8-bandit
beautifulsoup4==4.7.1
certifi==2019.6.16        # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
cohesion==0.9.1
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
cryptography==2.7         # via josepy, mozilla-django-oidc, pyopenssl
defusedxml==0.6.0         # via djangorestframework-xml
django-environ==0.4.5
django-grappelli==2.13.1
django-mailgun==0.9.1
django==2.2.2
djangorestframework-xml==1.4.0
djangorestframework==3.9.4
drf-dynamic-fields==0.3.1
entrypoints==0.3          # via flake8
flake8-bandit==2.1.1
flake8-blind-except==0.1.1
flake8-builtins==1.4.1
flake8-commas==2.0.0
flake8-comprehensions==2.1.0
flake8-debugger==3.1.0
flake8-import-order==0.18.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-pep3101==1.2.1
flake8-polyfill==1.0.2    # via flake8-bandit, pep8-naming
flake8-print==3.1.0
flake8-quotes==2.0.1
flake8-string-format==0.2.3
flake8-tidy-imports==2.0.0
flake8-todo==0.7
flake8==3.7.7
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11         # via bandit
idna==2.8                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10.1            # via coreschema
josepy==1.1.0             # via mozilla-django-oidc
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mozilla-django-oidc==1.2.2
pbr==5.3.1                # via stevedore
pep8-naming==0.8.2
pip-tools==3.8.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0        # via flake8, flake8-bandit, flake8-debugger, flake8-import-order, flake8-print, flake8-todo
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pyopenssl==19.0.0         # via josepy
pytz==2019.1              # via django
pyyaml==5.1.1
requests==2.22.0          # via coreapi, django-mailgun, mozilla-django-oidc
six==1.12.0               # via bandit, cryptography, django-mailgun, flake8-print, josepy, pip-tools, pyopenssl, stevedore
smmap2==2.0.5             # via gitdb2
soupsieve==1.9.2          # via beautifulsoup4
sqlparse==0.3.0           # via django
stevedore==1.30.1         # via bandit
uritemplate==3.0.0        # via coreapi
urllib3==1.25.3           # via requests
uwsgi==2.0.18

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via flake8-blind-except, flake8-import-order, josepy
